requirements.txt
pandas==2.1.4
openpyxl==3.1.2
SQLAlchemy==2.0.23
matplotlib==3.8.2
xlrd==2.0.1
mysql-connector-python==8.2.0
psycopg2==2.9.9
elasticsearch==8.0.1
requests==2.26.0